refactor hooks from store
