+++ ui/src/hooks/useNavigation.ts
@@ -1,0 +1,18 @@
+import NavigationState from "@/store/navigation";
+import { Actions } from "@/store/navigation/types";
+import { useRecoilState } from "recoil";
+
+export function useNavigation(): [number, Actions] {
+  const [value, setValue] = useRecoilState(NavigationState);
+
+  function change(_: React.SyntheticEvent<Element, Event>, value: number) {
+    setValue(value);
+  }
+
+  function reset() {
+    setValue(-1);
+  }
+
+  return [value, { change, reset }];
+}
+
