+++ ui/src/store/navigation/index.ts
@@ -1,24 +1,8 @@
-import { atom, useRecoilState } from 'recoil';
+import { atom } from 'recoil';
 
-import type { Actions } from './types';
-
-const navigationState = atom<number>({
-  key: 'navigation-state',
+const NavigationState = atom<number>({
+  key: 'NavigationState',
   default: 0,
 });
 
-function useNavigation(): [number, Actions] {
-  const [value, setValue] = useRecoilState(navigationState);
-
-  function change(_: React.SyntheticEvent<Element, Event>, value: number) {
-    setValue(value);
-  }
-
-  function reset() {
-    setValue(-1);
-  }
-
-  return [value, { change, reset }];
-}
-
-export default useNavigation;
+export default NavigationState;
