+++ ui/src/theme/Provider.tsx
@@ -17,7 +17,7 @@
 // SOFTWARE.
 import { ThemeProvider, createTheme } from '@mui/material/styles';
 
-import useTheme from '@/store/theme';
+import { useTheme } from '@/hooks/useTheme';
 
 import themes from './themes';
 import type { CustomThemeProviderProps } from './types';
